rename sun_model --> sun_equations and fix import into solar_array_calculator.py
