+++ solar_array_calculator.py
@@ -1,5 +1,5 @@
 """Calculates the power profile of the 'Polaris' solar car model.
-
+Jeremy Khoo, April 2017
 Given the latitude, longitude, GMT time offset, year, month, day, and angle that the car is facing relative to the 
 West direction, finds the power generated by the solar cells on the 'Polaris' solar car model at any time at that 
 location throughout the day.
@@ -8,7 +8,7 @@ location throughout the day.
 import sys
 import numpy as np
 import os.path
-import sunmodel as sm
+import sun_equations as sm
 
 
 class NodeVector:
